+++ projects/liars_poker_bot/src/cfragent.rs
@@ -69,7 +69,7 @@ impl<T: GameState, N: NodeStore<CFRNode>> CFRAgent<T, N> {
         return agent;
     }
 
-    fn get_policy(&mut self, istate: &IStateKey) -> ActionVec<f32> {
+    fn get_policy(&mut self, istate: &IStateKey) -> ActionVec<f64> {
         let n = self.ns.get(istate).unwrap();
         let p = n.borrow().get_average_strategy();
         self.ns.insert_node(*istate, n); // return the node
