+++ projects/liars_poker_bot/src/cfragent/cfrnode.rs
@@ -5,9 +5,9 @@ use crate::game::Action;
 /// Adapted from: https://towardsdatascience.com/counterfactual-regret-minimization-ff4204bf4205
 #[derive(Debug, Clone)]
 pub struct CFRNode {
-    pub regret_sum: ActionVec<f32>,
-    pub move_prob: ActionVec<f32>,
-    pub total_move_prob: ActionVec<f32>,
+    pub regret_sum: ActionVec<f64>,
+    pub move_prob: ActionVec<f64>,
+    pub total_move_prob: ActionVec<f64>,
 }
 
 impl CFRNode {
@@ -23,7 +23,7 @@ impl CFRNode {
     ///
     /// Defaults to a uniform action strategy if no regrets are present
     // Fix how this handles no data -- can't initialize all to 0
-    pub fn get_move_prob(&mut self, realization_weight: f32) -> ActionVec<f32> {
+    pub fn get_move_prob(&mut self, realization_weight: f64) -> ActionVec<f64> {
         let actions = &self.regret_sum.actions;
         let num_actions = actions.len();
         let mut normalizing_sum = 0.0;
@@ -37,7 +37,7 @@ impl CFRNode {
             if normalizing_sum > 0.0 {
                 self.move_prob[a] = self.move_prob[a] / normalizing_sum;
             } else {
-                self.move_prob[a] = 1.0 / num_actions as f32;
+                self.move_prob[a] = 1.0 / num_actions as f64;
             }
             self.total_move_prob[a] += realization_weight * self.move_prob[a];
         }
@@ -45,7 +45,7 @@ impl CFRNode {
         return self.move_prob.clone();
     }
 
-    pub fn get_average_strategy(&self) -> ActionVec<f32> {
+    pub fn get_average_strategy(&self) -> ActionVec<f64> {
         let actions = &self.regret_sum.actions;
 
         let mut avg_strat = ActionVec::new_from(&actions);
@@ -58,7 +58,7 @@ impl CFRNode {
             if normalizing_sum > 0.0 {
                 avg_strat[a] = self.total_move_prob[a] / normalizing_sum;
             } else {
-                avg_strat[a] = 1.0 / self.move_prob.len() as f32;
+                avg_strat[a] = 1.0 / self.move_prob.len() as f64;
             }
         }
